+++ user-service/src/main/java/com/gaebal_easy/client/user/presentation/UserController.java
@@ -1,7 +1,7 @@
 package com.gaebal_easy.client.user.presentation;
 
 import com.gaebal_easy.client.user.application.service.JoinService;
-import com.gaebal_easy.client.user.presentation.dtos.JoinRequest;
+import com.gaebal_easy.client.user.presentation.dto.JoinRequest;
 import gaebal_easy.common.global.dto.ApiResponseData;
 import gaebal_easy.common.global.enums.Role;
 import gaebal_easy.common.global.utils.RequiredRole;
