+++ user-service/src/main/java/com/gaebal_easy/client/user/presentation/dto/JoinRequest.java
@@ -1,4 +1,4 @@
-package com.gaebal_easy.client.user.presentation.dtos;
+package com.gaebal_easy.client.user.presentation.dto;
 
 import gaebal_easy.common.global.enums.Role;
 import lombok.Getter;
