+++ user-service/src/main/java/com/gaebal_easy/client/user/presentation/dto/UserRequest.java
@@ -1,0 +1,4 @@
+package com.gaebal_easy.client.user.presentation.dto;
+
+public class UserRequest {
+}
