+++ user-service/src/main/java/com/gaebal_easy/client/user/presentation/dto/UserResponse.java
@@ -1,0 +1,4 @@
+package com.gaebal_easy.client.user.presentation.dto;
+
+public class UserResponse {
+}
