--- user-service/src/main/java/com/gaebal_easy/client/user/presentation/dtos/UserRequest.java
@@ -1,4 +1,0 @@
-package com.gaebal_easy.client.user.presentation.dtos;
-
-public class UserRequest {
-}
